fix(application,adapters,ui): address PR review feedback
- Surface poolLookupFailures count in ListSupportedPositionsResult
  so the controller can warn about incomplete data instead of silently
  returning empty results during transient outages
- Move DTO validators (isPositionSummaryDto, isPositionDetailDto,
  isPositionSummaryRecord, isPositionSummaryDtoArray) from apps/app
  to packages/application/src/dto/validation.ts to respect the
  architecture boundary (apps/app must be composition-only)
- Fix stale tick-oriented fallback in PositionDetailPresenter:
  currentPriceLabel fallback now uses price-space format,
  label fallbacks derive from numeric bound prices instead of
  empty strings
