+++ packages/application/src/use-cases/positions/ListSupportedPositions.ts
@@ -7,6 +7,7 @@ import { buildPositionDisplayBounds } from './buildPositionDisplayBounds.js';
 export type ListSupportedPositionsResult = {
   positions: LiquidityPosition[];
   summaryDtos: PositionSummaryDto[];
+  poolLookupFailures: number;
 };
 
 export async function listSupportedPositions(params: {
@@ -18,12 +19,13 @@ export async function listSupportedPositions(params: {
   const uniquePoolIds = [...new Set(positions.map((p) => p.poolId))];
   const poolDataMap = new Map<PoolId, Awaited<ReturnType<SupportedPositionReadPort['getPoolData']>>>();
 
+  let poolLookupFailures = 0;
   await Promise.allSettled(uniquePoolIds.map(async (poolId) => {
     try {
       const poolData = await params.positionReadPort.getPoolData(poolId);
       if (poolData) poolDataMap.set(poolId, poolData);
     } catch {
-      // individual pool lookup failure — positions in this pool will be excluded
+      poolLookupFailures++;
     }
   }));
 
@@ -71,5 +73,5 @@ export async function listSupportedPositions(params: {
     });
   }
 
-  return { positions, summaryDtos };
+  return { positions, summaryDtos, poolLookupFailures };
 }
