+++ packages/ui/src/presenters/PositionDetailPresenter.ts
@@ -14,7 +14,7 @@ function normalizePositionDetailDto(dto: Partial<PositionDetailDto> & Pick<Posit
     poolId,
     tokenPairLabel: dto.tokenPairLabel ?? `Pool ${poolId}`,
     currentPrice: dto.currentPrice ?? 0,
-    currentPriceLabel: dto.currentPriceLabel ?? `tick: ${dto.currentPrice ?? 0}`,
+    currentPriceLabel: dto.currentPriceLabel ?? `${dto.currentPrice ?? 0}`,
     feeRateLabel: dto.feeRateLabel ?? '',
     rangeState: dto.rangeState ?? 'in-range',
     rangeDistance: dto.rangeDistance ?? { belowLowerPercent: 0, aboveUpperPercent: 0 },
@@ -22,8 +22,8 @@ function normalizePositionDetailDto(dto: Partial<PositionDetailDto> & Pick<Posit
     monitoringStatus: dto.monitoringStatus ?? 'active',
     lowerBoundPrice: dto.lowerBoundPrice ?? 0,
     upperBoundPrice: dto.upperBoundPrice ?? 0,
-    lowerBoundLabel: dto.lowerBoundLabel ?? '',
-    upperBoundLabel: dto.upperBoundLabel ?? '',
+    lowerBoundLabel: dto.lowerBoundLabel ?? `${dto.lowerBoundPrice ?? 0}`,
+    upperBoundLabel: dto.upperBoundLabel ?? `${dto.upperBoundPrice ?? 0}`,
     sqrtPrice: dto.sqrtPrice ?? '0',
     unclaimedFees: dto.unclaimedFees ?? {
       feeOwedA: { ...EMPTY_TOKEN_AMOUNT },
